Abstract adapter tests to reduce code duplication
